Crate ethnum

This crate implements 256-bit integer types.

The implementation tries to follow as closely as possible to primitive integer types, and should implement all the common methods and traits as the primitive integer types.

Modules

intrinsics

This module contains intrinsics used by the I256 and U256 implementations.

prelude

Convenience re-export of 256-integer types and as- conversion traits.

Macros

int

Macro for 256-bit signed integer literal.

uint

Macro for 256-bit unsigned integer literal.

Structs

I256

A 256-bit signed integer type.

U256

A 256-bit unsigned integer type.

Traits

AsI256

This trait defines as conversions (casting) from primitive types to I256.

AsU256

This trait defines as conversions (casting) from primitive types to U256.

Type Aliases

i256

A 256-bit signed integer type.

u256

A 256-bit unsigned integer type.