ethnum

Module intrinsics

This module contains intrinsics used by the I256 and U256 implementations.

Stability

Be careful when using these intrinsics directly. Semantic versioning API compatibility is not guaranteed for any of these intrinsics.

Modules

signed

Module containing signed wrapping functions around various intrinsics that are agnostic to sign.

Functions

add2

add3

ctlz

cttz

iaddc

idiv2

idiv3

idivmod4

imulc

irem2

irem3

isubc

mul2

mul3

rol3

ror3

sar2

sar3

shl2

shl3

shr2

shr3

sub2

sub3

uaddc

udiv2

udiv3

udivmod4

umulc

umulddi3

urem2

urem3

usubc