ethnum::intrinsics

Module signed

Module containing signed wrapping functions around various intrinsics that are agnostic to sign.

This module can be helpful when using intrinsics directly.

Re-exports

pub use super::add2 as uadd2;

pub use super::add3 as uadd3;

pub use super::ctlz as uctlz;

pub use super::cttz as ucttz;

pub use super::iaddc;

pub use super::idiv2;

pub use super::idiv3;

pub use super::imulc;

pub use super::irem2;

pub use super::irem3;

pub use super::isubc;

pub use super::mul2 as umul2;

pub use super::mul3 as umul3;

pub use super::rol3 as urol3;

pub use super::ror3 as uror3;

pub use super::sar2 as isar2;

pub use super::sar3 as isar3;

pub use super::shl2 as ushl2;

pub use super::shl3 as ushl3;

pub use super::shr2 as ushr2;

pub use super::shr3 as ushr3;

pub use super::sub2 as usub2;

pub use super::sub3 as usub3;

pub use super::uaddc;

pub use super::udiv2;

pub use super::udiv3;

pub use super::umulc;

pub use super::urem2;

pub use super::urem3;

pub use super::usubc;

Functions

iadd2

iadd3

ictlz

icttz

imul2

imul3

irol3

iror3

ishl2

ishl3

isub2

isub3