Module ethnum::intrinsics::signed

Module containing signed wrapping functions around various intrinsics that are agnostic to sign.

This module can be helpful when using intrinsics directly.

Functions

iadd2

iadd3

iaddc

ictlz

icttz

idiv2

idiv3

imul2

imul3

imulc

irem2

irem3

irol3

iror3

isar2

isar3

ishl2

ishl3

isub2

isub3

isubc

uadd2

uadd3

uaddc

uctlz

ucttz

udiv2

udiv3

umul2

umul3

umulc

urem2

urem3

urol3

uror3

ushl2

ushl3

ushr2

ushr3

usub2

usub3

usubc